+++ packages/cli/bin/index.js
@@ -3,6 +3,7 @@ const { cli, console: { printError } } = require('../dist');
 
 cli(process.cwd(), process.argv)
   .catch(error => {
+    // console.log(error);
     printError(error.message);
     process.exit(1);
   })